+++ tracee-ebpf/main.go
@@ -16,7 +16,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/aquasecurity/tracee/tracee"
+	"github.com/aquasecurity/tracee/tracee-ebpf/tracee"
 	"github.com/syndtr/gocapability/capability"
 	"github.com/urfave/cli/v2"
 )
