+++ tracee-rules/engine/engine_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
+	tracee "github.com/aquasecurity/tracee/tracee-ebpf/tracee/external"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
-	tracee "github.com/aquasecurity/tracee/tracee/external"
 	"github.com/stretchr/testify/assert"
 )
