+++ tracee-rules/go.mod
@@ -3,8 +3,11 @@ module github.com/aquasecurity/tracee/tracee-rules
 go 1.15
 
 require (
-	github.com/aquasecurity/tracee v0.4.0
+	github.com/aquasecurity/tracee/tracee-ebpf v0.4.0
 	github.com/open-policy-agent/opa v0.25.2
 	github.com/stretchr/testify v1.5.1
 	github.com/urfave/cli/v2 v2.3.0
 )
+
+//TODO: remove this before release
+replace github.com/aquasecurity/tracee/tracee-ebpf => github.com/aquasecurity/tracee v0.4.0
