+++ tracee-rules/input.go
@@ -11,8 +11,8 @@ import (
 	"os"
 	"strings"
 
+	tracee "github.com/aquasecurity/tracee/tracee-ebpf/tracee/external"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
-	tracee "github.com/aquasecurity/tracee/tracee/external"
 )
 
 var errHelp = errors.New("user has requested help text")
