+++ tracee-rules/signatures/golang/examples/example.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 	"strings"
 
+	tracee "github.com/aquasecurity/tracee/tracee-ebpf/tracee/external"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
-	tracee "github.com/aquasecurity/tracee/tracee/external"
 )
 
 // counter is a simple demo signature that counts towards a target
