+++ tracee-rules/signatures/golang/stdio_over_socket.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
+	tracee "github.com/aquasecurity/tracee/tracee-ebpf/tracee/external"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
-	tracee "github.com/aquasecurity/tracee/tracee/external"
 )
 
 type stdioOverSocket struct {
