+++ tracee-rules/signatures/golang/stdio_over_socket_test.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	tracee "github.com/aquasecurity/tracee/tracee/external"
+	tracee "github.com/aquasecurity/tracee/tracee-ebpf/tracee/external"
 	"testing"
 
 	"github.com/aquasecurity/tracee/tracee-rules/signatures/signaturestest"
