+++ tracee-rules/signatures/rego/regosig/traceerego_test.go
@@ -5,9 +5,9 @@ import (
 	"reflect"
 	"testing"
 
+	tracee "github.com/aquasecurity/tracee/tracee-ebpf/tracee/external"
 	"github.com/aquasecurity/tracee/tracee-rules/signatures/signaturestest"
 	"github.com/aquasecurity/tracee/tracee-rules/types"
-	tracee "github.com/aquasecurity/tracee/tracee/external"
 )
 
 func TestGetMetadata(t *testing.T) {
